refactor: combined ModifierHelper into ModifierEventHandler. Pulled out CollectEquipmentSlotsEvent.
